backend/requirements.txt
# NexHost V6 - ClawCloud Ready
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP & AI
httpx==0.26.0
aiohttp==3.9.1

# Security
python-dotenv==1.0.0
cryptography==42.0.0

# Bot libraries (pre-installed)
requests==2.31.0
python-telegram-bot==20.7
pyTelegramBotAPI==4.16.0
aiogram==3.3.0
schedule==1.2.1
APScheduler==3.10.4
colorama==0.4.6
rich==13.7.0
loguru==0.7.2
pycryptodome==3.20.0
PyYAML==6.0.1
